% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scores.R
\docType{class}
\name{AbstractConditionalScore-class}
\alias{AbstractConditionalScore-class}
\alias{evaluate,AbstractConditionalScore,TwoStageDesign-method}
\title{[ToDO]}
\usage{
\S4method{evaluate}{AbstractConditionalScore,TwoStageDesign}(s, design, x1,
  ...)
}
\arguments{
\item{s}{score}

\item{design}{design}

\item{x1}{stage-one outcome}

\item{...}{further optional arguments}
}
\description{
[ToDo]
}
