% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AffineScoreFunction.R
\docType{class}
\name{AffineScore-class}
\alias{AffineScore-class}
\alias{AffineScore}
\alias{evaluate,AffineScore,TwoStageDesign-method}
\title{Affine functions of scores}
\usage{
AffineScore(scores, coefs, intercept)

\S4method{evaluate}{AffineScore,TwoStageDesign}(s, design, ...)
}
\arguments{
\item{scores}{cf. corresponding slot}

\item{coefs}{cf. corresponding slot}

\item{intercept}{cf. corresponding slot}

\item{s}{score}

\item{design}{design}

\item{...}{further optional arguments}
}
\description{
These methods allow to add and multiplicate two or several
conditional resp. unconditional scores.
They can also be weighted by numeric coefficients and shifted
by a numeric intercept.
This yields to new scores as sum of scores and is therefore a useful
tool to combine scores.
}
\section{Slots}{

\describe{
\item{\code{scores}}{the list of scores}

\item{\code{coefs}}{numeric vector of the same length as \code{scores}, holding the
coefficients}

\item{\code{intercept}}{the intercept for the affine function

\code{AffineScore} allows to add scores of arbitrary class and shift
them by the \code{intercept}.
\code{AffineUnconditionalScore} allows the same for scores of class
\code{UnconditionalScore} and \code{AffineConditionalScore} for
scores of class \code{ConditionalScore}.
By the methods \code{+} and \code{*} scores can be added and multiplicated,
respectively.
Note that it is not possible to use these methods for a mixture of
\code{ConditionalScore} and \code{UnconditionalScore} as these
require different evaluation techniques.
However, for both score classes multiplication and addition with
numerics is provided.}
}}

