% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDistribution.R
\docType{class}
\name{ANOVA-class}
\alias{ANOVA-class}
\alias{ANOVA}
\title{Analysis of Variance}
\usage{
ANOVA(n_groups)
}
\arguments{
\item{n_groups}{number of groups to be compared}
}
\description{
ANOVA is used to test whether there is a significant difference between the means of groups.
The sample size which \code{adoptr} returns is the total sample size,
}
\examples{
model <- ANOVA(3L)

}
\seealso{
see \code{\link{probability_density_function}} and
   \code{\link{cumulative_distribution_function}} to evaluate the pdf
   and the cdf, respectively.
}
