% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularization.R
\docType{class}
\name{AverageN2-class}
\alias{AverageN2-class}
\alias{AverageN2}
\alias{evaluate,AverageN2,TwoStageDesign-method}
\title{Regularization via L1 norm}
\usage{
AverageN2()

\S4method{evaluate}{AverageN2,TwoStageDesign}(s, design, specific = TRUE,
  ...)
}
\arguments{
\item{s}{an \code{RegularizationScore}}

\item{design}{a \code{TwoStageDesign}}

\item{specific}{logical, flag for switching to design-specific implementation.}

\item{...}{further optimal arguments}
}
\description{
Implements the L1-norm of the design's stage-two sample size function.
}
\section{Methods (by generic)}{
\itemize{
\item \code{evaluate}: generic implementation of evaluating a smoothness
score. Uses adaptive Gaussian quadrature for integration and might be
more efficiently implemented by specific \code{TwoStageDesign}-classes
(cf. \code{\link{.evaluate}}).
}}

