% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDistribution.R
\docType{class}
\name{Binomial-class}
\alias{Binomial-class}
\alias{Binomial}
\alias{quantile,Binomial-method}
\title{Binomial data distribution}
\usage{
Binomial(rate_control, two_armed = TRUE)

\S4method{quantile}{Binomial}(x, probs, n, theta, ...)
}
\arguments{
\item{rate_control}{cf. slot 'rate_control'}

\item{two_armed}{logical indicating if a two-armed trial is regarded}

\item{x}{outcome}

\item{probs}{vector of probabilities}

\item{n}{sample size}

\item{theta}{distribution parameter}

\item{...}{further optional arguments}
}
\description{
Implements the normal approximation for a test on rates.
The reponse rate in the control group,
\ifelse{html}{\out{r<sub>C</sub>}}{\eqn{r_C}}, has to be specified by
\code{rate_control}.
The null hypothesis is:
\ifelse{html}{\out{r<sub>E</sub> &le; r<sub>E</sub>}}{\eqn{r_E <= r_C}},
where \ifelse{html}{\out{r<sub>E</sub>}}{\eqn{r_E}} denotes the response rate
in the invervention group.
It is tested versus the alternative
\ifelse{html}{\out{r<sub>E</sub> > r<sub>E</sub>}}{\eqn{r_E > r_C}}
}
\details{
All priors have to be defined for the rate difference
\ifelse{html}{\out{r<sub>E</sub> - r<sub>C</sub>}}{\eqn{r_E - r_C}}.

\code{Binomial} contains the standard deviation under the null hypothesis
to save runtime.
}
\section{Slots}{

\describe{
\item{\code{rate_control}}{assumed response rate in control group}

\item{\code{sigma_0}}{standard deviation of rate difference under the null hypothesis}
}}

\examples{
datadist <- Binomial(rate_control = 0.2, two_armed = FALSE)

}
\seealso{
see \code{\link{probability_density_function}} and
   \code{\link{cumulative_distribution_function}} to evaluate the pdf
   and the cdf, respectively.
}
