% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChiSquaredDistribution.R
\docType{class}
\name{ChiSquared-class}
\alias{ChiSquared-class}
\title{Chi-Squared data distribution}
\arguments{
\item{x}{outcome}

\item{n}{sample size}

\item{theta}{distribution parameter}

\item{...}{further optional arguments}
}
\description{
Viel Spaß, Nico.
Implements the normal approximation for a test on rates.
The reponse rate in the control group,
\ifelse{html}{\out{r<sub>C</sub>}}{\eqn{r_C}}, has to be specified by
\code{rate_control}.
The null hypothesis is:
\ifelse{html}{\out{r<sub>E</sub> &le; r<sub>C</sub>}}{\eqn{r_E <= r_C}},
where \ifelse{html}{\out{r<sub>E</sub>}}{\eqn{r_E}} denotes the response rate
in the invervention group.
It is tested against the alternative
\ifelse{html}{\out{r<sub>E</sub> > r<sub>C</sub>}}{\eqn{r_E > r_C}}.
The test statistic is given as
\ifelse{html}{\out{X<sub>1</sub> = &radic;n (r<sub>E</sub> - r<sub>C</sub>) / &radic;(2  r<sub>0</sub> (1-r<sub>0</sub>))}}{\eqn{X_1 = \sqrt{n}(r_E - r_C) / \sqrt{2 r_0 (1- r_0)}}},
where \ifelse{html}{\out{r<sub>0</sub>}}{\eqn{r_0}} denotes the mean between
\ifelse{html}{\out{r<sub>E</sub>}}{\eqn{r_E}} and
\ifelse{html}{\out{r<sub>C</sub>}}{\eqn{r_C}} in the two-armed case,
and \ifelse{html}{\out{r<sub>E</sub>}}{\eqn{r_E}} in the one-armed case.#'
All priors have to be defined for the rate difference
\ifelse{html}{\out{r<sub>E</sub> - r<sub>C</sub>}}{\eqn{r_E - r_C}}.
}
\section{Slots}{

\describe{
\item{\code{df}}{cf. parameter 'df'}

\item{\code{ncp}}{cf. parameter 'ncp'}
}}

