% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConditionalSampleSize.R
\docType{class}
\name{ConditionalSampleSize-class}
\alias{ConditionalSampleSize-class}
\alias{ConditionalSampleSize}
\alias{evaluate,ConditionalSampleSize,Design-method}
\title{Conditional sample size of a design given stage-one outcome}
\usage{
ConditionalSampleSize(dist, prior)

\S4method{evaluate}{ConditionalSampleSize,Design}(s, design, x1, ...)
}
\arguments{
\item{dist}{data distribution}

\item{prior}{prior distribution}

\item{s}{conditional score object to evaluate}

\item{design}{a \code{Design}}

\item{x1}{stage one outcome (note that n1 is available from \code{design})}

\item{...}{further optional parameters}
}
\description{
This score simply evaluates \code{n(d, x1)}.
The data distribution and prior are only relevant when it is integrated.
}
\details{
See documentation in \link{ConditionalScore-class}
    for further details and inherited methods.
}
\section{Functions}{
\itemize{
\item \code{ConditionalSampleSize}: constructor
}}

\section{Slots}{

\describe{
\item{\code{distribution}}{\code{DataDistribution} object specifying the data distribution
given the parameter \eqn{\theta} of the model}

\item{\code{prior}}{\code{Prior} object specifying the data distribution
given the parameter \eqn{\theta} of the model}
}}

