% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scores.R
\docType{class}
\name{ConditionalScore-class}
\alias{ConditionalScore-class}
\alias{expected,ConditionalScore-method}
\title{Class for conditional scoring function}
\usage{
\S4method{expected}{ConditionalScore}(s, ...)
}
\arguments{
\item{s}{conditional score object to evaluate}

\item{...}{optional arguments}
}
\description{
[ToDo]
}
\section{Methods (by generic)}{
\itemize{
\item \code{expected}: integrate a \code{ConditionalScore} over the
stage-one outcome; return object of class \code{\link{IntegralScore-class}}.
}}

