% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\docType{class}
\name{ConstraintsCollection-class}
\alias{ConstraintsCollection-class}
\alias{evaluate,ConstraintsCollection,Design,ANY-method}
\alias{subject_to}
\title{Collection of constraints}
\usage{
\S4method{evaluate}{ConstraintsCollection,Design,ANY}(s, design, x1, ...)

subject_to(...)
}
\arguments{
\item{...}{arbitrary number of (un)conditional constraints}
}
\description{
\code{subject_to(...)} can be used to generate an object of class
     \code{ConstrintsCollection} from an arbitrary number of (un)conditional
     constraints.
}
\section{Slots}{

\describe{
\item{\code{unconditional_constraints}}{[todo]}

\item{\code{conditional_constraints}}{[todo]}
}}

