% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scores.R
\name{.evaluate}
\alias{.evaluate}
\title{Design specific implementation for evaluating integral scores}
\usage{
.evaluate(s, design, ...)
}
\arguments{
\item{s}{score object}

\item{design}{design object}

\item{...}{further optional arguments}
}
\description{
This generic is not exported and for internal use only.
It allows \code{\link{evaluate}} which is implemented generically via
adaptive Gaussian quadrature to selectively dispatch to a design-specific
implementation.
This is mostly exploited during optiization to allow more efficient and
stable evaluation of integral scores.
}
\keyword{internal}
