% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scores.R
\docType{methods}
\name{expected}
\alias{expected}
\alias{expected,ConditionalScore-method}
\title{Compute the expectation of a conditional score}
\usage{
expected(s, ...)

\S4method{expected}{ConditionalScore}(s, ...)
}
\arguments{
\item{s}{ConditionalScore}

\item{...}{further optional arguments}
}
\value{
an object of class \code{\link{IntegralScore}}
}
\description{
By the method \code{expected} any \code{\link{ConditionalScore}}
can be integrated over the full \eqn{x_1}-range and returns an
\code{\link{IntegralScore}}. I.e., for a conditional score
\eqn{s(design, x_1)}{s(design, x1)} the integral
\eqn{\int s(design, x_1) d x_1}{over s(design, x1)}
is computed.
}
\examples{
# creates power under point mass prior
expected(ConditionalPower(Normal(), PointMassPrior(.3, 1)))

}
