% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{integrate_rule}
\alias{integrate_rule}
\title{Integration via the Gauss-Legendre quadrature}
\usage{
integrate_rule(f, low, up, x, weights)
}
\arguments{
\item{f}{function to be integrated}

\item{low}{lower bound of integration}

\item{up}{upper bound of integration}

\item{x}{normalized integration knots on [-1, 1]}

\item{weights}{integration weights}
}
\description{
Integration via the Gauss-Legendre quadrature
}
\keyword{internal}
