% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{GaussLegendreRule}
\alias{GaussLegendreRule}
\title{Compute weights an knots for Gauss-Legendre integration}
\usage{
GaussLegendreRule(order)
}
\arguments{
\item{order}{Number of nodes to be used}
}
\description{
Implementation of Gauss Legendre integration rule and computation of integral
based on implementation at https://rdrr.io/cran/SMR/src/R/GaussLegendre.R
\code{GaussLegendreRule} gives the nodes and weights for the interval \eqn{(-1, 1)}.
}
\keyword{internal}
