% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gq_design.R
\docType{methods}
\name{gq_design}
\alias{gq_design}
\alias{gq_design,numeric-method}
\title{Design with Gaussian quadrature integration rule}
\usage{
gq_design(n1, c1f, c1e, n2, c2, order = 5, ...)

\S4method{gq_design}{numeric}(n1, c1f, c1e, n2, c2, order = 5, ...)
}
\arguments{
\item{n1}{stage-one sample size}

\item{c1f}{early stopping for futility boundary}

\item{c1e}{early stopping for efficacy boundary}

\item{n2}{second-stage sample size. vector for adaptive and
integer for group-sequential design}

\item{c2}{secnod-stage decision boundary}

\item{order}{order (i.e. number of pivot points in the interior of [c1f, c1e])
of the Gaussian quadrature rule to use for integration}

\item{...}{futher arguments}
}
\description{
gq_design simply creates a TwoStageDesign object with a Gaussian
    quadrature numerical integration rule. If length(n2) = 1 and
    length(c2) > 1 an object of class \code{GSDesign} is created.
}
\section{Methods (by class)}{
\itemize{
\item \code{numeric}: Method to create design with Gaussian quadrature rule
and Legendre nodes
}}

