% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GQDesign.R
\docType{class}
\name{GQDesign-class}
\alias{GQDesign-class}
\alias{GQDesign}
\alias{update,GQDesign-method}
\alias{n1,GQDesign-method}
\alias{get_knots,GQDesign-method}
\alias{n2,GQDesign,numeric-method}
\alias{c2,GQDesign,numeric-method}
\alias{as.numeric,GQDesign-method}
\title{Gaussian-quadrature based implementation of \code{Design}}
\usage{
GQDesign(n1, c1f, c1e, n2_pivots, c2_pivots, order)

\S4method{update}{GQDesign}(object, params, ...)

\S4method{n1}{GQDesign}(d, ...)

\S4method{get_knots}{GQDesign}(d, ...)

\S4method{n2}{GQDesign,numeric}(d, x1, ...)

\S4method{c2}{GQDesign,numeric}(d, x1, ...)

\S4method{as.numeric}{GQDesign}(x, ...)
}
\arguments{
\item{n1}{stage-one sample size}

\item{c1f}{early stopping for futility boundary}

\item{c1e}{early stopping for efficacy boundary}

\item{n2_pivots}{vector of length order giving the values of n2 at the
pivot points of the Gaussian quadrature rule [TODO: these are not available during construction]}

\item{c2_pivots}{vector of length order giving the values of c2 at the
pivot points of the Gaussian quadrature rule [TODO: these are not available during construction]}

\item{order}{order (i.e. number of pivot points in the interior of [c1f, c1e])
of the Gaussian quadrature rule to use for integration}

\item{object}{GQDesign to update with given parameter values.}

\item{d}{object of class GQDesign}

\item{x1}{stage-one outcome}

\item{x}{GQDesign}
}
\description{
Implements a class for two-stage designs where n2 and c2 are approximated
by linear interpolation of pivot points and integration is handled via
Gaussian quadrature rules.
The pivots are chosen to be the evaluation points of the selected
Gaussian quadrature rule and guarantee a stable evaluation during
optimization.

[TODO: implement a constructor taking in n2 and c2 functions?]
}
\section{Methods (by generic)}{
\itemize{
\item \code{update}: return a new design object with parameters given by a
single numeric vector of length 2*order + 3, inverse of \code{as.numeric(design)}.

\item \code{n1}: stage-one sample size

\item \code{get_knots}: get the pivots points (knots) of the Gaussian quadrature
rule.

\item \code{n2}: stage-two sample size by linear interpolation of n2_pivots

\item \code{c2}: stage-two critical value by linear interpolation of n2_pivots

\item \code{as.numeric}: convert GQDesign to vector of tunable parameters via
\code{c(x@n1, x@c1f, x@c1e, x@n2_pivots, x@c2_pivots)}; inverse of update();
}}

\section{Slots}{

\describe{
\item{\code{rule}}{data.frame, the integration rule, result of a call to
\code{gaussquad::legendre.quadrature.rules(order)[[order]])} during
construction}
}}

