% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupSequentialDesign.R
\docType{class}
\name{GroupSequentialDesignSurvival-class}
\alias{GroupSequentialDesignSurvival-class}
\title{Group-sequential two-stage designs for time-to-event-endpoints}
\description{
Group-sequential designs for time-to-event-endpoints are a subclass of both
\code{TwoStageDesignSurvival} and \code{GroupSequentialDesign}.
}
\seealso{
\code{\link{TwoStageDesignSurvival-class}} and \code{\link{GroupSequentialDesign}}
for superclasses and inherited methods.
}
