% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupSequentialDesign.R
\name{gs2ts}
\alias{gs2ts}
\title{Convert a group-sequential design to a two-stage design}
\usage{
gs2ts(d)
}
\arguments{
\item{d}{object of class \code{GSDesign}}
}
\description{
Convert a group-sequential design to a two-stage design
}
