% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scores.R
\docType{class}
\name{IntegralScore-class}
\alias{IntegralScore-class}
\alias{evaluate,IntegralScore,Design,ANY-method}
\title{Unconditional score class obtained by integration of a \code{ConditionalScore}}
\usage{
\S4method{evaluate}{IntegralScore,Design,ANY}(s, design, x1 = NA,
  specific = TRUE, ..)
}
\arguments{
\item{s}{an \code{IntegralScore}}

\item{design}{a \code{Design}}

\item{x1}{required argument for generic, unused}

\item{specific}{logical, flag for switching to design-specific implementation.}

\item{...}{further optimal arguments}
}
\description{
Unconditional score class obtained by integration of a \code{ConditionalScore}
}
\section{Methods (by generic)}{
\itemize{
\item \code{evaluate}: generic implementation of evaluating an integral
score. Uses adaptive Gaussian quadrature for integration and might be
more efficiently implemented by specific \code{Design}-classes
(cf. \code{\link{.evaluate}}).
}}

\section{Slots}{

\describe{
\item{\code{cs}}{the underlying \code{ConditionalScore}}
}}

