% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survival.R
\docType{class}
\name{Survival-class}
\alias{Survival-class}
\alias{Survival}
\alias{probability_density_function,Survival,numeric,numeric,numeric-method}
\alias{cumulative_distribution_function,Survival,numeric,numeric,numeric-method}
\title{Log-rank test}
\usage{
Survival(event_rate, two_armed = TRUE)

\S4method{probability_density_function}{Survival,numeric,numeric,numeric}(dist, x, n, theta, ...)

\S4method{cumulative_distribution_function}{Survival,numeric,numeric,numeric}(dist, x, n, theta, ...)
}
\arguments{
\item{event_rate}{probability that a subject will eventually have an event}

\item{two_armed}{logical indicating if a two-armed trial is regarded}

\item{x}{outcome}

\item{n}{sample size}

\item{theta}{distribution parameter}

\item{...}{further optional arguments}
}
\description{
Implements the normal approximation of the log-rank test statistic.
}
\examples{
datadist <- Survival(event_rate=0.6, two_armed=TRUE)

}
\seealso{
see \code{\link{probability_density_function}} and
   \code{\link{cumulative_distribution_function}} to evaluate the pdf
   and the cdf, respectively.
}
