% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimize.R
\name{minimize}
\alias{minimize}
\title{Find optimal two-stage design by constraint minimization}
\usage{
minimize(objective, subject_to, initial_design, lower_boundary_design,
  upper_boundary_design, c2_monotone = FALSE, post_process = FALSE,
  opts = list(algorithm = "NLOPT_LN_COBYLA", xtol_rel = 1e-05, maxeval =
  10000), ...)
}
\arguments{
\item{objective}{objective function}

\item{subject_to}{constraint collection}

\item{initial_design}{initial guess (x0 for nloptr)}

\item{lower_boundary_design}{design specifying the lower boundary}

\item{upper_boundary_design}{design specifying the upper boundary}

\item{c2_monotone}{should the c2-function be forced to be monotoneously decreasing?}

\item{post_process}{should the sample sizes be integers?}

\item{opts}{options list passed to nloptr}

\item{...}{further optional arguments passed to \code{\link{nloptr}}}
}
\value{
\item{design}{ The resulting optimal design}
        \item{details}{ A list with the following arguments
        \itemize{
        \item{design}{ The optimal design without rounded n-values}
        \item{design_post}{ The optimal design after post-processing.
        Returns \code{NULL} if post_process was set to \code{FALSE}.}
        \item{nloptr_return}{ The return of the nloptr call for the
        non-post-processed design.}
        \item{nloptr_return_post}{ The return of the nloptr call for the
        post-processed design.
        Returns \code{NULL} if post_process was set to \code{FALSE}.}
        }
        }
}
\description{
\code{minimize} takes an unconditioonal score
and a constraint set (or single constraint) of conditional and/or unconditional
scores and solves the corresponding constraint minimization problem
using \code{nloptr} (using COBYLA by default).
}
