% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularization.R
\docType{class}
\name{N1-class}
\alias{N1-class}
\alias{N1}
\alias{evaluate,N1,TwoStageDesign-method}
\title{Regularize n1}
\usage{
N1()

\S4method{evaluate}{N1,TwoStageDesign}(s, design, optimization = FALSE,
  ...)
}
\arguments{
\item{s}{an object of class \code{N1}}

\item{design}{the design to compute the smoothness term for}

\item{...}{further optional arguments}
}
\description{
\code{N1} is a class that penalizes the \code{n1} value of
a design.
}
