% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDistribution.R
\docType{class}
\name{Normal-class}
\alias{Normal-class}
\alias{Normal}
\alias{probability_density_function,Normal,numeric,numeric,numeric-method}
\alias{cumulative_distribution_function,Normal,numeric,numeric,numeric-method}
\alias{quantile,DataDistribution-method}
\title{Normal data distribution}
\usage{
Normal()


  \S4method{probability_density_function}{Normal,numeric,numeric,numeric}(dist,
  x, n, theta, ...)


  \S4method{cumulative_distribution_function}{Normal,numeric,numeric,numeric}(dist,
  x, n, theta, ...)

\S4method{quantile}{DataDistribution}(x, probs, n, theta, ...)
}
\arguments{
\item{x}{observed z value}

\item{n}{stage size}

\item{probs}{vector of probablities}
}
\description{
Implements a normal data distribution for z-values given an observed z-value
and stage size.
Standard deviation is 1 and mean \eqn{\theta\sqrt n} where
\eqn{\theta} is the standardized effect size.
}
\section{Methods (by generic)}{
\itemize{
\item \code{probability_density_function}: PDF

\item \code{cumulative_distribution_function}: CDF

\item \code{quantile}: quantile function
}}

