% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneStageDesign.R
\docType{class}
\name{OneStageDesignSurvival-class}
\alias{OneStageDesignSurvival-class}
\title{One-stage designs for time-to-event endpoints}
\description{
\code{OneStageDesignSurvival} is a subclass of both \code{OneStageDesign} and
\code{TwoStageDesignSurvival}.
}
\seealso{
\code{\link{TwoStageDesignSurvival-class}} and \code{\link{OneStageDesign-class}}
for superclasses and inherited methods.
}
