% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otsd.R
\docType{package}
\name{otsd}
\alias{otsd}
\alias{otsd-package}
\title{Optimal Two-Stage Designs}
\description{
This package provides functionality to compute and analyze optimal
two-stage designs for two-stage superiority tests based on a Z statistic.
A range of pre-implemented objective/constraint functions are provided.
}
\section{Quickstart}{


For a sample workflow and a quick demo of the capabilities, see the
vignette [todo: link].
}

\section{Designs}{


list designs
}

\section{Prios}{


List prior classes
}

\section{Scores}{


list score classes
}

