\name{print.adoptrOptimizationResult}
\alias{print.adoptrOptimizationResult}
\alias{print}
\title{Printing an optimization result}
\usage{
print(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{further arguments passed form other methods}
}
\description{
Printing an optimization result
}
