% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoStageDesign.R
\name{print.TwoStageDesignSummary}
\alias{print.TwoStageDesignSummary}
\title{Print obejct of class TwoStageDesignSummary}
\usage{
\method{print}{TwoStageDesignSummary}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{unused}
}
\description{
Print obejct of class TwoStageDesignSummary
}
