% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoStageDesign.R
\docType{methods}
\name{show,TwoStageDesign-method}
\alias{show,TwoStageDesign-method}
\title{Show method for TwoStageDesign objects}
\usage{
\S4method{show}{TwoStageDesign}(object)
}
\arguments{
\item{object}{design to show}
}
\description{
Only states the class itself. More information is given by the respective
\code{summary} method, cf. \link{TwoStageDesign-class}.
}
