% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smoothness_n2.R
\docType{class}
\name{Smoothness_n2-class}
\alias{Smoothness_n2-class}
\alias{Smoothness_n2}
\alias{evaluate,Smoothness_n2,Design-method}
\title{Quadratic smoothness penalty term}
\usage{
Smoothness_n2(h = sqrt(.Machine$double.eps))

\S4method{evaluate}{Smoothness_n2,Design}(s, design, specific = TRUE,
  ...)
}
\arguments{
\item{h}{positive number, see slot \code{h}}

\item{s}{an object of class \code{Smoothness_n2}}

\item{design}{the design to compute the smoothness term for}

\item{specific}{logical, should a design-specific implementation be used?
defaults to \code{TRUE}. If \code{TRUE}, looks for design-specific
\code{.evaluate} method and calls it}

\item{...}{further optional arguments}
}
\description{
\code{Smoothness_n2} is a generic class for implementing a smoothness penalty
via the average squared first derivative of the stage two sample size function
\code{n2} of a two-stage design.
The only parameter is the width used for the finite differences, \code{h}.
The generic implementation only evluates \code{n2} in the interior of the
continuation region of a design.
}
\section{Slots}{

\describe{
\item{\code{h}}{positive number giving the width of the central finite difference
interval for approximating the first derivative.}
}}

