% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scores.R
\docType{class}
\name{SmoothnessL1-class}
\alias{SmoothnessL1-class}
\alias{SmoothnessL1}
\alias{evaluate,SmoothnessL1,TwoStageDesign-method}
\title{Smoothness via L1 norm}
\usage{
SmoothnessL1()

\S4method{evaluate}{SmoothnessL1,TwoStageDesign}(s, design,
  specific = TRUE, ...)
}
\arguments{
\item{specific}{logical, flag for switching to design-specific implementation.}

\item{...}{further optimal arguments}
}
\description{
Implements the smoothness of a design by computing the
L1-norm of the stage-two sample size function.
}
\section{Methods (by generic)}{
\itemize{
\item \code{evaluate}: generic implementation of evaluating a smoothness
score. Uses adaptive Gaussian quadrature for integration and might be
more efficiently implemented by specific \code{TwoStageDesign}-classes
(cf. \code{\link{.evaluate}}).
}}

