% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularization.R
\docType{class}
\name{SmoothnessN2-class}
\alias{SmoothnessN2-class}
\alias{SmothnessN2}
\alias{SmoothnessN2}
\title{Quadratic smoothness penalty term}
\usage{
SmoothnessN2(h = 0.1)
}
\arguments{
\item{h}{positive number, see slot \code{h}}
}
\value{
an object of class \code{\link{SmoothnessN2}}
}
\description{
\code{SmoothnessN2} is a generic class for implementing a smoothness penalty
via the average squared second derivative of the stage two sample size
function \eqn{n_2}{n2} of a two-stage design.
The only parameter is the width used for the finite differences, \eqn{h}{h}.
The generic implementation only evaluates \eqn{n_2}{n2} in the interior of
the continuation region of a design.
}
\section{Slots}{

\describe{
\item{\code{h}}{positive number giving the width of the central finite difference
interval for approximating the second derivative.}
}}

