% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scores.R
\docType{class}
\name{SmoothnessScore-class}
\alias{SmoothnessScore-class}
\alias{evaluate,SmoothnessScore,TwoStageDesign-method}
\title{Smoothness of a design}
\usage{
\S4method{evaluate}{SmoothnessScore,TwoStageDesign}(s, design, ...)
}
\description{
Unconditional score class to evaluate the smoothness of
a design's sample size function
}
\section{Methods (by generic)}{
\itemize{
\item \code{evaluate}: not implemented, just raises a 'not implemented'
}}

