% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudentDistribution.R
\docType{class}
\name{Student-class}
\alias{Student-class}
\alias{Student}
\alias{quantile,Student-method}
\alias{simulate,Student,numeric-method}
\title{Student's t data distribution}
\usage{
Student(two_armed = TRUE)

\S4method{quantile}{Student}(x, probs, n, theta, ...)

\S4method{simulate}{Student,numeric}(object, nsim, n, theta, seed = NULL, ...)
}
\arguments{
\item{two_armed}{logical indicating if a two-armed trial is regarded}

\item{x}{outcome}

\item{probs}{vector of probabilities}

\item{n}{sample size}

\item{theta}{distribution parameter}

\item{...}{further optional arguments}

\item{object}{object of class \code{Student}}

\item{nsim}{number of simulation runs}

\item{seed}{random seed}
}
\description{
Implements exact t-distributions instead of a normal approximation
}
\examples{
datadist <- Student(two_armed = TRUE)

}
\seealso{
see \code{\link{probability_density_function}} and
   \code{\link{cumulative_distribution_function}} to evaluate the pdf
   and the cdf, respectively.
}
