% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoStageDesign.R
\docType{methods}
\name{summary,TwoStageDesign-method}
\alias{summary,TwoStageDesign-method}
\title{Summarize TwoStageDesign objects with optional set of scores}
\usage{
\S4method{summary}{TwoStageDesign}(object, ..., rounded = T)
}
\arguments{
\item{object}{design object to summarize}

\item{...}{optinal additional named UnconditionalScores}

\item{rounded}{should rounded n-values be used?}
}
\description{
[TODO]
}
