% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoStageDesign.R, R/OneStageDesign.R,
%   R/GroupSequentialDesign.R
\name{SurvivalDesign}
\alias{SurvivalDesign}
\alias{SurvivalDesign,TwoStageDesign-method}
\alias{SurvivalDesign,OneStageDesign-method}
\alias{SurvivalDesign,GroupSequentialDesign-method}
\title{SurvivalDesign}
\usage{
SurvivalDesign(design, ...)

\S4method{SurvivalDesign}{TwoStageDesign}(design, event_rate)

\S4method{SurvivalDesign}{OneStageDesign}(design, event_rate)

\S4method{SurvivalDesign}{GroupSequentialDesign}(design, event_rate)
}
\arguments{
\item{design}{design that should be converted to a survival design}

\item{event_rate}{probability that a subject in either group will eventually have an event}
}
\description{
\code{SurvivalDesign} is a function that converts an arbitrary design to a survival design.
}
\examples{
design <- get_initial_design(0.4,0.025,0.1)
SurvivalDesign(design,0.8)

}
