% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoStageDesign.R
\docType{methods}
\name{plot,TwoStageDesign-method}
\alias{plot,TwoStageDesign-method}
\title{Plot TwoStageDesign with optional set of conditional scores}
\usage{
\S4method{plot}{TwoStageDesign}(x, y = NULL, ..., k = 100)
}
\arguments{
\item{x}{design to plot}

\item{y}{not used}

\item{...}{optional named (Un)ConditionalScores to plot / include in the summary}

\item{k}{number of points to use for plotting}
}
\description{
[TODO]
}
