% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scores.R
\docType{class}
\name{UnconditionalScore-class}
\alias{UnconditionalScore-class}
\alias{evaluate,UnconditionalScore,TwoStageDesign-method}
\title{Abstract class for unconditional scoring function}
\usage{
\S4method{evaluate}{UnconditionalScore,TwoStageDesign}(s, design, ...)
}
\arguments{
\item{s}{an \code{IntegralScore}}

\item{design}{a \code{TwoStageDesign}}

\item{...}{further optional arguments}
}
\description{
[ToDo]
}
