% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scores.R
\name{working-with-scores}
\alias{working-with-scores}
\alias{evaluate}
\alias{integrate}
\title{Working with scores}
\usage{
evaluate(s, design, ...)

integrate(s, ...)
}
\arguments{
\item{s}{score}

\item{design}{\code{TwoStageDesign} object}

\item{...}{further optional arguments}
}
\description{
Both \code{\link{ConditionalScore-class}} as well as
\code{\link{UnconditionalScore-class}} implement \code{evaluate} methods
which handle the actual computation of the score given a design and
(for conditional scores) an interim result.
Conditional scores additionally implement an \code{integrate} method
to obtain the corresponding unconditional \code{\link{IntegralScore-class}}.
}
