% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChiSquaredDistribution.R
\name{ZSquared}
\alias{ZSquared}
\alias{get_tau_ZSquared}
\title{Squared normally distributed random variable}
\usage{
ZSquared(two_armed = TRUE)

get_tau_ZSquared(mu, sigma = 1)
}
\arguments{
\item{two_armed}{logical indicating if a two-aremd trial is regarded}

\item{mu}{mean of Z}

\item{sigma}{standard deviation of Z}
}
\description{
Implementation of \eqn{Z^2}, where \eqn{Z} is normally distributed with mean \eqn{\mu} and variance
\eqn{\sigma^2}. \eqn{Z^2} is chi-squared distributed with \eqn{1} degree of freedom and non-centrality parameter \eqn{(\mu/\sigma)^2}.
The function \code{get_tau_ZSquared} computes the factor \eqn{\tau=(\mu/\sigma)^2}, such that
\eqn{\tau} is the equivalent of \eqn{\theta} in the normally distributed case.
}
\examples{
zsquared <- ZSquared(FALSE)

H1 <- PointMassPrior(get_tau_ZSquared(0.4, 1)

}
\seealso{
see \code{\link{probability_density_function}} and
   \code{\link{cumulative_distribution_function}} to evaluate the pdf
   and the cdf, respectively.
}
