% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adoptr.R
\docType{package}
\name{adoptr}
\alias{adoptr}
\alias{adoptr-package}
\title{Adaptive Optimal Two-Stage Designs}
\description{
The \pkg{adoptr} package provides functionality to explore custim optimal
two-stage designs for one or two-arm superiority tests.
Currently, only (asymptotically) normal test statistics are supported.
}
\section{Quickstart}{


For a sample workflow and a quick demo of the capabilities, see
\href{here}{https://kkmann.github.io/adoptr/articles/adoptr.html}.
}

\section{Designs}{


\pkg{adoptr} currently supports \code{\link{TwoStageDesign}},
\code{\link{GroupSequentialDesign}}, and \code{\link{OneStageDesign}}.
}

\section{Data distributions}{


[Todo]
}

\section{Priors}{


Both \code{\link{ContinuousPrior}} and \code{\link{PointMassPrior}} are
supported for the single parameter of a \code{\link{DataDistribution}}.
}

\section{Scores}{


TODO: list score classes
}

