% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AffineScoreFunction.R
\docType{class}
\name{AffineScore-class}
\alias{AffineScore-class}
\alias{AffineScore}
\alias{evaluate,AffineScore,Design-method}
\title{Affine functions of scores}
\usage{
AffineScore(scores, coefs, intercept)

\S4method{evaluate}{AffineScore,Design}(s, design, ...)
}
\arguments{
\item{scores}{cf. corresponding slot}

\item{coefs}{cf. corresponding slot}

\item{intercept}{cf. corresponding slot}

\item{s}{score}

\item{design}{design}

\item{...}{further optional arguments}
}
\description{
[TODO]
}
\section{Slots}{

\describe{
\item{\code{scores}}{the list of scores}

\item{\code{coefs}}{numeric vector of the same length as \code{scores}, holding the
coefficients}

\item{\code{intercept}}{the intercept for the affine function}
}}

