% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularization.R
\docType{class}
\name{AverageN2-class}
\alias{AverageN2-class}
\alias{AverageN2}
\alias{evaluate,AverageN2,TwoStageDesign-method}
\title{Regularization via L1 norm}
\usage{
AverageN2()

\S4method{evaluate}{AverageN2,TwoStageDesign}(s, design,
  optimization = FALSE, subdivisions = 10000L, ...)
}
\arguments{
\item{s}{a score of class \code{AverageN2}}

\item{design}{a \code{TwoStageDesign}}

\item{...}{further optimal arguments}

\item{specific}{logical, flag for switching to design-specific implementation.}
}
\description{
Implements the L1-norm of the design's stage-two sample size function.
The average of the stage-two sample size without weighting with
the data distribution is computed.
This can be interpreted as integration over a unifrom prior on
the continuation region.
Adding the L1-norm with a small regularization parameter can make
the optimization more stable.
}
\section{Methods (by generic)}{
\itemize{
\item \code{evaluate}: generic implementation of evaluating a smoothness
score. Uses adaptive Gaussian quadrature for integration and might be
more efficiently implemented by specific \code{TwoStageDesign}-classes.
}}

