% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularization.R
\docType{class}
\name{AverageN2-class}
\alias{AverageN2-class}
\alias{AverageN2}
\title{Regularization via L1 norm}
\usage{
AverageN2()
}
\value{
an object of class \code{\link{AverageN2}}
}
\description{
Implements the L1-norm of the design's stage-two sample size function.
The average of the stage-two sample size without weighting with
the data distribution is computed.
This can be interpreted as integration over a unifrom prior on
the continuation region.
}
\examples{
avn2 <- AverageN2()

}
\seealso{
\code{\link{N1}} for penalizing n1 values
}
