% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_designs.R
\docType{methods}
\name{get_lower_boundary_design}
\alias{get_lower_boundary_design}
\alias{get_upper_boundary_design}
\alias{get_lower_boundary_design,OneStageDesign-method}
\alias{get_lower_boundary_design,GroupSequentialDesign-method}
\alias{get_lower_boundary_design,TwoStageDesign-method}
\alias{get_upper_boundary_design,OneStageDesign-method}
\alias{get_upper_boundary_design,GroupSequentialDesign-method}
\alias{get_upper_boundary_design,TwoStageDesign-method}
\title{Boundary designs}
\usage{
get_lower_boundary_design(initial_design, ...)

get_upper_boundary_design(initial_design, ...)

\S4method{get_lower_boundary_design}{OneStageDesign}(initial_design, ...)


  \S4method{get_lower_boundary_design}{GroupSequentialDesign}(initial_design,
  ...)

\S4method{get_lower_boundary_design}{TwoStageDesign}(initial_design, ...)

\S4method{get_upper_boundary_design}{OneStageDesign}(initial_design, ...)


  \S4method{get_upper_boundary_design}{GroupSequentialDesign}(initial_design,
  ...)

\S4method{get_upper_boundary_design}{TwoStageDesign}(initial_design, ...)
}
\arguments{
\item{initial_design}{The initial design}

\item{...}{optimal arguments}
}
\description{
Compute boundary designs. These can be used for optimization and are
implemented by default in \link{minimize}.
}
