% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{methods}
\name{boundary-designs}
\alias{boundary-designs}
\alias{get_lower_boundary_design,UnconditionalScore,ConstraintsCollection,OneStageDesign-method}
\alias{get_lower_boundary_design,UnconditionalScore,ConstraintsCollection,GroupSequentialDesign-method}
\alias{get_lower_boundary_design,UnconditionalScore,ConstraintsCollection,TwoStageDesign-method}
\alias{get_upper_boundary_design}
\alias{get_upper_boundary_design,UnconditionalScore,ConstraintsCollection,OneStageDesign-method}
\alias{get_upper_boundary_design,UnconditionalScore,ConstraintsCollection,GroupSequentialDesign-method}
\alias{get_upper_boundary_design,UnconditionalScore,ConstraintsCollection,TwoStageDesign-method}
\title{Boundary designs}
\usage{
get_lower_boundary_design(objective, subject_to, initial_design, ...)


  \S4method{get_lower_boundary_design}{UnconditionalScore,ConstraintsCollection,OneStageDesign}(objective,
  subject_to, initial_design, ...)


  \S4method{get_lower_boundary_design}{UnconditionalScore,ConstraintsCollection,GroupSequentialDesign}(objective,
  subject_to, initial_design, ...)


  \S4method{get_lower_boundary_design}{UnconditionalScore,ConstraintsCollection,TwoStageDesign}(objective,
  subject_to, initial_design, ...)

get_upper_boundary_design(objective, subject_to, initial_design, ...)


  \S4method{get_upper_boundary_design}{UnconditionalScore,ConstraintsCollection,OneStageDesign}(objective,
  subject_to, initial_design, ...)


  \S4method{get_upper_boundary_design}{UnconditionalScore,ConstraintsCollection,GroupSequentialDesign}(objective,
  subject_to, initial_design, ...)


  \S4method{get_upper_boundary_design}{UnconditionalScore,ConstraintsCollection,TwoStageDesign}(objective,
  subject_to, initial_design, ...)
}
\arguments{
\item{objective}{An object of class \code{\link{UnconditionalScore-class}} to be minimized}

\item{subject_to}{A \code{\link{ConstraintsCollection-class}} defining the constraints}

\item{initial_design}{The initial design}
}
\description{
Compute boundary designs. These can be used for optimization and are
implemented by default in \link{minimize}.
}
