% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChiSquaredDistribution.R, R/FDistribution.R
\docType{class}
\name{ChiSquared-class}
\alias{ChiSquared-class}
\alias{Pearson2xK-class}
\alias{ZSquared-class}
\alias{ChiSquared}
\alias{quantile,ChiSquared-method}
\alias{simulate,ChiSquared,numeric-method}
\alias{ANOVA}
\title{Chi-Squared data distribution}
\usage{
ChiSquared(df, multiplier = 1)

\S4method{quantile}{ChiSquared}(x, probs, n, theta, ...)

\S4method{simulate}{ChiSquared,numeric}(object, nsim, n, theta, seed = NULL, ...)

ANOVA(n_groups)
}
\arguments{
\item{df}{number of degrees of freedom}

\item{probs}{vector of probabilities}

\item{object}{object of class \code{Binomial}}

\item{nsim}{number of simulation runs}

\item{seed}{random seed}

\item{rate_control}{assumed response rate in control group}

\item{two_armed}{logical indicating if a two-armed trial is regarded}
}
\description{
Implements a chi-squared distribution. The classes \code{Pearson2xk}
and \code{ZSquared} are subclasses, used in two different situations.
\code{Pearson2xk} can be used in trials with binary endpoints
when testing k groups for homogeneity in response rates.
\code{ZSquared} simply implements the square of a normally distributed random variable
with mean \eqn{\mu} and standard deviation \eqn{\sigma^2}.

This test is used in trials with binary endpoints
when testing k groups for homogeneity in response rates. The null hypothesis is
\ifelse{html}{\out{r<sub>1</sub>}=...=\out{r<sub>k</sub>}}{\eqn{r_1=...=r_k}}, and the
alternative is that there exists a pair of groups with differing rates.
}
\details{
Note that \code{simulate} for class \code{ChiSquared} simulates the
   normal approximation of the test statistic.
}
\examples{
datadist <- ChiSqured(df=4)

datadist <- Binomial(rate_control = 0.2, two_armed = FALSE)

}
\seealso{
see \code{\link{probability_density_function}} and
   \code{\link{cumulative_distribution_function}} to evaluate the pdf
   and the cdf, respectively.

see \code{\link{probability_density_function}} and
   \code{\link{cumulative_distribution_function}} to evaluate the pdf
   and the cdf, respectively.
}
