% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConditionalPower.R
\docType{class}
\name{ConditionalPower-class}
\alias{ConditionalPower-class}
\alias{ConditionalPower}
\alias{evaluate,ConditionalPower,Design-method}
\title{Conditional power of a design given stage-one outcome}
\usage{
ConditionalPower(dist, prior)

\S4method{evaluate}{ConditionalPower,Design}(s, design, x1, ...)
}
\arguments{
\item{dist}{data distribution}

\item{prior}{prior distribution}

\item{s}{conditional score object to evaluate}

\item{design}{a \code{Design}}

\item{x1}{stage one outcome (note that n1 is available from \code{design})}

\item{...}{further optional parameters}
}
\description{
This score evaluates to \eqn{P[X_2 > c2(design, X_1)|X_1 = x_1]}.
Note that the distribution of \eqn{X_2} is the posterior predictive after
observing \eqn{X_1=x_1}.
}
\details{
See documentation in \link{ConditionalScore-class}
    for further details and inherited methods.
}
\section{Functions}{
\itemize{
\item \code{ConditionalPower}: constructor
}}

\section{Slots}{

\describe{
\item{\code{distribution}}{\code{DataDistribution} object specifying the data distribution
given the parameter \eqn{\theta} of the model}

\item{\code{prior}}{\code{Prior} object specifying the data distribution
given the parameter \eqn{\theta} of the model}
}}

