% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scores.R
\docType{class}
\name{ConditionalScore-class}
\alias{ConditionalScore-class}
\alias{evaluate,ConditionalScore,Design-method}
\alias{integrate,ConditionalScore-method}
\title{Abstract class for conditional scoring function}
\usage{
\S4method{evaluate}{ConditionalScore,Design}(s, design, x1, ...)

\S4method{integrate}{ConditionalScore}(s, ...)
}
\arguments{
\item{s}{an \code{ConditionalScore}}

\item{design}{a \code{Design}}

\item{x1}{stage one outcome (note that n1 is available from \code{design})}

\item{s}{an \code{ConditionalScore}}

\item{design}{a \code{Design}}

\item{x1}{stage one outcome (note that n1 is available from \code{design})}
}
\description{
[ToDo]
}
\section{Methods (by generic)}{
\itemize{
\item \code{evaluate}: bla

\item \code{integrate}: Integrate a Conditional Score over the stage-one outcome
}}

