% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\docType{class}
\name{Constraint-class}
\alias{Constraint-class}
\alias{evaluate,Constraint,TwoStageDesign-method}
\alias{ConditionalConstraint-class}
\alias{<=,AbstractConditionalScore,numeric-method}
\alias{>=,AbstractConditionalScore,numeric-method}
\alias{UnconditionalConstraint-class}
\alias{<=,UnconditionalScore,numeric-method}
\alias{>=,UnconditionalScore,numeric-method}
\title{Formulating constraints}
\usage{
\S4method{evaluate}{Constraint,TwoStageDesign}(s, design, ...)

\S4method{<=}{AbstractConditionalScore,numeric}(e1, e2)

\S4method{>=}{AbstractConditionalScore,numeric}(e1, e2)

\S4method{<=}{UnconditionalScore,numeric}(e1, e2)

\S4method{>=}{UnconditionalScore,numeric}(e1, e2)
}
\arguments{
\item{s}{constraint to evaluate}

\item{design}{TwoStageDesign to evaluate}

\item{...}{further optional arguments}

\item{e1}{first comparator}

\item{e2}{second comparator}
}
\description{
Conceptually, constraints work very similar to scores (any score can be put in
a constraint).
Currently, only constraints of the form 'score <=/>= x' are admissable.#'
}
\details{
[TODO: currently we only support scores on the LHS - this can be easily extended!]

[TODO: do we also want to support scores vs scores comparisons?]

[TODO: conditional constraints are only evaluated at the continuation area]
}
\examples{
cp          <- ConditionalPower(Normal, PointPrior(0.4))
pow         <- integrate(cp)
constraint1 <- pow >= 0.8 # an unconditional power constraint
constraint2 <- cp >= 0.7 # a conditional power constraint

}
