% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\docType{class}
\name{Constraint-class}
\alias{Constraint-class}
\alias{evaluate,Constraint,Design-method}
\alias{ConditionalConstraint-class}
\alias{<=,AbstractConditionalScore,numeric-method}
\alias{>=,AbstractConditionalScore,numeric-method}
\alias{UnconditionalConstraint-class}
\alias{<=,UnconditionalScore,numeric-method}
\alias{>=,UnconditionalScore,numeric-method}
\title{Formulating constraints}
\usage{
\S4method{evaluate}{Constraint,Design}(s, design, ...)

\S4method{<=}{AbstractConditionalScore,numeric}(e1, e2)

\S4method{>=}{AbstractConditionalScore,numeric}(e1, e2)

\S4method{<=}{UnconditionalScore,numeric}(e1, e2)

\S4method{>=}{UnconditionalScore,numeric}(e1, e2)
}
\arguments{
\item{s}{constraint to evaluate [TODO make naming of arguments for evaluate more generic!]}

\item{design}{design to evaluate}

\item{...}{further optional arguments}

\item{e1}{first comparator}

\item{e2}{second comparator}
}
\description{
[TODO]
}
\details{
[TODO: currently we only support scores on the LHS - this can be easily extended!]

[TODO: do we also want to support scores vs scores comparisons?]
}
