% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\docType{class}
\name{ConstraintsCollection-class}
\alias{ConstraintsCollection-class}
\alias{ConstraintsCollection}
\title{Collection of constraints}
\arguments{
\item{s}{constraint collection}

\item{design}{design}
}
\description{
Collection of constraints
}
\section{Slots}{

\describe{
\item{\code{unconditional_constraints}}{a list of elements of class \code{UnconditionalConstraint}}

\item{\code{conditional_constraints}}{a list of elements of class \code{ConditionalConstraint}

A \code{ConstraintsCollection} is a collection of unconditional and
conditional constraints. In order to evaluate these correctly, they
have to be defined in two different slots.}
}}

\seealso{
A \code{ConstraintsCollection} can be created by \code{\link{subject_to}}.
}
