% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDistribution.R
\docType{class}
\name{DataDistribution-class}
\alias{DataDistribution-class}
\alias{quantile,DataDistribution-method}
\alias{probability_density_function}
\alias{probability_density_function,DataDistribution,numeric,numeric,numeric-method}
\alias{cumulative_distribution_function}
\alias{cumulative_distribution_function,DataDistribution,numeric,numeric,numeric-method}
\title{Data distributions}
\usage{
\S4method{quantile}{DataDistribution}(dist, x, probs, n, theta, ...)

probability_density_function(dist, x, n, theta, ...)


  \S4method{probability_density_function}{DataDistribution,numeric,numeric,numeric}(dist,
  x, n, theta, ...)

cumulative_distribution_function(dist, x, n, theta, ...)


  \S4method{cumulative_distribution_function}{DataDistribution,numeric,numeric,numeric}(dist,
  x, n, theta, ...)
}
\arguments{
\item{dist}{data distribution object}

\item{x}{outcome}

\item{probs}{numeric vector of probabilities}

\item{n}{sample size}

\item{theta}{distribution parameter}

\item{...}{further optional arguments}
}
\description{
\code{DataDistribution} is an abstract class used to represent the distribution
of a sufficient statistic \code{x} given a sample size \code{n} and a
single parameter value \code{theta}.
}
\details{
This abstraction layer allows representation of t-distributions
(unknown variance), normal distribution (known variance), and normal
approximation of a binary endpoint.
Currently, only the normal case is implemented with \code{\link{Normal-class}}.

The logical option \code{two_armed} allows to decide whether a one-arm or
a two-arm (the default) design should be computed. In the case of a two-arm
design the function \code{\link{plot}} and \code{\link{summary}} describe
the sample size per group.
}
\section{Methods (by generic)}{
\itemize{
\item \code{quantile}: quantile function of the respective distribution.

\item \code{probability_density_function}: probability density function given outcome,
sample size and parameter; must be implemented.

\item \code{cumulative_distribution_function}: cumulative distribution function given outcome,
sample size and parameter; must be implemented.
}}

\section{Slots}{

\describe{
\item{\code{two_armed}}{Should a two-armed design be used?}
}}

