% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDistribution.R
\docType{class}
\name{DataDistribution-class}
\alias{DataDistribution-class}
\alias{probability_density_function}
\alias{cumulative_distribution_function}
\alias{show,DataDistribution-method}
\title{Data distributions}
\usage{
probability_density_function(dist, x, n, theta, ...)

cumulative_distribution_function(dist, x, n, theta, ...)

\S4method{show}{DataDistribution}(object)
}
\arguments{
\item{dist}{data distribution object}

\item{x}{outcome}

\item{n}{sample size}

\item{theta}{distribution parameter}

\item{...}{further optional arguments}

\item{object}{data distribution to show}
}
\description{
\code{DataDistribution} is an abstract class used to represent the distribution
of a sufficient statistic \code{x} given a sample size \code{n} and a
single parameter value \code{theta}.

Only states the class itself.
}
\details{
This abstraction layer allows representation of t-distributions
(unknown variance), normal distribution (known variance), and normal
approximation of a binary endpoint.
Currently, only the normal case is implemented with \code{\link{Normal-class}}.

The logical option \code{two_armed} allows to decide whether a one-arm or
a two-arm (the default) design should be computed. In the case of a two-arm
design all sample sizes are per group.
}
\section{Slots}{

\describe{
\item{\code{two_armed}}{Should a two-armed design be used?}
}}

