% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scores.R
\docType{methods}
\name{expected}
\alias{expected}
\alias{expected,ConditionalScore-method}
\title{Compute the expectation of a conditional score}
\usage{
expected(s, ...)

\S4method{expected}{ConditionalScore}(s, ...)
}
\arguments{
\item{s}{ConditionalScore}

\item{...}{further optional arguments}
}
\value{
an object of class \code{\link{IntegralScore}}
}
\description{
TODO
}
\examples{
expected(ConditionalPower(Normal(), PointMassPrior(.3, 1))
# creates power under point mass prior

}
