% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{GaussLegendreIntegral}
\alias{GaussLegendreIntegral}
\title{Integration via the Gauss Legendre quadrature}
\usage{
GaussLegendreIntegral(f, low, up, rule)
}
\arguments{
\item{f}{function to be integrated}

\item{low}{lower bound of integration}

\item{up}{upper bound of integration}

\item{rule}{integration rule as computed by \code{GaussLegendreRule}}
}
\description{
Integration via the Gauss Legendre quadrature
}
