% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GQDesign.R
\docType{class}
\name{GQDesign-class}
\alias{GQDesign-class}
\alias{GQDesign}
\alias{update,GQDesign-method}
\alias{get_knots}
\alias{get_knots,GQDesign-method}
\alias{n2,GQDesign,numeric-method}
\alias{c2,GQDesign,numeric-method}
\alias{as.numeric,GQDesign-method}
\title{Gaussian-quadrature based implementation of \code{Design}}
\usage{
GQDesign(n1, c1f, c1e, n2_pivots, c2_pivots, order)

\S4method{update}{GQDesign}(object, params, ...)

get_knots(d, ...)

\S4method{get_knots}{GQDesign}(d, ...)

\S4method{n2}{GQDesign,numeric}(d, x1, ...)

\S4method{c2}{GQDesign,numeric}(d, x1, ...)

\S4method{as.numeric}{GQDesign}(x, ...)
}
\arguments{
\item{n1}{stage-one sample size}

\item{c1f}{early stopping for futility boundary}

\item{c1e}{early stopping for efficacy boundary}

\item{n2_pivots}{vector of length order giving the values of n2 at the
pivot points of the Gaussian quadrature rule [TODO: these are not available during construction]}

\item{c2_pivots}{vector of length order giving the values of c2 at the
pivot points of the Gaussian quadrature rule [TODO: these are not available during construction]}

\item{order}{order (i.e. number of pivot points in the interior of [c1f, c1e])
of the Gaussian quadrature rule to use for integration}

\item{object}{\code{Design} object to update with new design parameter values}

\item{params}{vector of design parameters (must be in same order as returned
by \code{as.numeric(design)})}

\item{...}{further optional arguments}

\item{d}{\code{Design} object}

\item{x1}{stage-one outcome}

\item{x}{\code{Design} object to conver to numeric vector of design parameters.}
}
\description{
Implements a class for two-stage designs where n2 and c2 are approximated
by linear interpolation of pivot points and integration is handled via
Gaussian quadrature rules.
The pivots are chosen to be the evaluation points of the selected
Gaussian quadrature rule and guarantee a stable evaluation during
optimization.
See \code{\link{Design-class}} for details on the inherited methods.
}
\section{Methods (by generic)}{
\itemize{
\item \code{get_knots}: get the pivots points (knots) of the Gaussian quadrature
rule.
}}

\section{Slots}{

\describe{
\item{\code{rule}}{data.frame, the integration rule, result of a call to
\code{\link{GaussLegendreRule}} during
construction}

\item{\code{n1}}{cf. parameter}

\item{\code{c1f}}{cf. parameter}

\item{\code{c1e}}{cf. parameter}

\item{\code{n2_pivots}}{cf. parameter}

\item{\code{c2_pivots}}{cf. parameter}
}}

