% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupSequentialDesign.R
\docType{class}
\name{GroupSequentialDesign-class}
\alias{GroupSequentialDesign-class}
\alias{GroupSequentialDesign}
\alias{n2,GroupSequentialDesign,numeric-method}
\alias{TwoStageDesign,GroupSequentialDesign-method}
\alias{show,GroupSequentialDesign-method}
\title{Group-sequential design}
\usage{
GroupSequentialDesign(n1, c1f, c1e, n2_pivots, c2_pivots, order = NULL,
  ...)

\S4method{n2}{GroupSequentialDesign,numeric}(d, x1, round = TRUE, ...)

\S4method{TwoStageDesign}{GroupSequentialDesign}(d, tunable, ...)

\S4method{show}{GroupSequentialDesign}(object)
}
\arguments{
\item{n1}{cf. slot}

\item{c1f}{cf. slot}

\item{c1e}{cf. slot}

\item{n2_pivots}{stage-two sample size}

\item{c2_pivots}{stage-two rejection boundary}

\item{order}{order of Gaussian quadrature}

\item{...}{further optional arguments}

\item{d}{object of class \code{GSDesign}}

\item{x1}{stage-one outcome}

\item{round}{logical, should integer sample size or real sample size be
returned?}

\item{tunable}{c.f. slot}

\item{object}{design to show}
}
\value{
an object of class \code{GroupSequentialDesign}
}
\description{
The class \code{GroupSequentialDesign} allows the implementation
and usage of group-sequential designs. These are two-stage designs
with a constant stage-two sample size function.
Therefore, the length of \code{n2_pivots} is required to be 1 in this
class.
\code{GroupSequentialDesign} is a subclass of \link{TwoStageDesign}.

If \code{order} is not defined, the length of the parameter \code{c2_pivots}
defines the order of the Gaussian quadrature integration rule.

Only states the class itself.
}
\section{Slots}{

\describe{
\item{\code{n1}}{stage-one sample size}

\item{\code{c1f}}{early stopping for futility boundary}

\item{\code{c1e}}{early stopping for efficacy boundary}

\item{\code{n2_pivots}}{stage-two sample size upon continuation}

\item{\code{c2_pivots}}{vector of length order giving the values of c2 at the
pivot points of the numeric integration rule [TODO: these are not available during construction]}

\item{\code{x1_norm_pivots}}{normalized pivots for integration rule (in [-1, 1])}

\item{\code{weights}}{weights of conditional score values at x1_norm_pivots for
approximating the integral over x1.}

\item{\code{tunable}}{named logical vector indicating whether corresponding slot is considered a tunable parameter}
}}

