% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupSequentialDesign.R
\docType{class}
\name{GSDesign-class}
\alias{GSDesign-class}
\alias{GSDesign}
\alias{tunable_parameters,GSDesign-method}
\alias{update,GSDesign-method}
\alias{n2,GSDesign,numeric-method}
\alias{TwoStageDesign,GSDesign-method}
\title{Group-sequential design}
\usage{
GSDesign(n1, c1f, c1e, n2_continue, c2_pivots, x1_norm_pivots, weights)

\S4method{tunable_parameters}{GSDesign}(x, ...)

\S4method{update}{GSDesign}(object, params, ...)

\S4method{n2}{GSDesign,numeric}(d, x1, ...)

\S4method{TwoStageDesign}{GSDesign}(s, ...)
}
\arguments{
\item{n1}{cf. slot}

\item{c1f}{cf. slot}

\item{c1e}{cf. slot}

\item{n2_continue}{cf. slot}

\item{c2_pivots}{cf. slot}

\item{x1_norm_pivots}{cf. slot}

\item{weights}{cf. slot}

\item{x}{object to get parameters from}

\item{...}{further optional arguments}

\item{object}{object to update}

\item{params}{vector of design parameters (must be in same order as returned
by \code{as.numeric(design)})}

\item{d}{design object}

\item{x1}{stage-one outcome}

\item{s}{object of class \code{GSDesign}}
}
\description{
[ToDo]
}
\section{Slots}{

\describe{
\item{\code{n1}}{stage-one sample size}

\item{\code{c1f}}{early stopping for futility boundary}

\item{\code{c1e}}{early stopping for efficacy boundary}

\item{\code{n2_continue}}{stage-two sample size}

\item{\code{c2_pivots}}{vector of length order giving the values of c2 at the
pivot points of the numeric integration rule [TODO: these are not available during construction]}

\item{\code{x1_norm_pivots}}{normalized pivots for integration rule (in [-1, 1])}

\item{\code{weights}}{weights of conditional score values at x1_norm_pivots for
approximating the integral over x1.}
}}

