% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupSequentialDesign.R, R/regularization.R
\docType{class}
\name{GSDesign-class}
\alias{GSDesign-class}
\alias{GSDesign}
\alias{n2,GSDesign,numeric-method}
\alias{TwoStageDesign,GSDesign-method}
\alias{evaluate,SmoothnessN2,GSDesign-method}
\title{Group-sequential design}
\usage{
GSDesign(n1, c1f, c1e, n2_pivots, c2_pivots, x1_norm_pivots, weights)

\S4method{n2}{GSDesign,numeric}(d, x1, round = TRUE, ...)

\S4method{TwoStageDesign}{GSDesign}(d, tunable, ...)

\S4method{evaluate}{SmoothnessN2,GSDesign}(s, design, ...)
}
\arguments{
\item{n1}{cf. slot}

\item{c1f}{cf. slot}

\item{c1e}{cf. slot}

\item{n2_pivots}{cf. slot}

\item{c2_pivots}{cf. slot}

\item{x1_norm_pivots}{cf. slot}

\item{weights}{cf. slot}

\item{d}{object of class \code{GSDesign}}

\item{x1}{stage-one outcome}

\item{round}{logical, should integer sample size or real sample size be
returned?}

\item{...}{further optional arguments}

\item{tunable}{c.f. slot}

\item{s}{an object of class \code{SmoothnessN2}}

\item{design}{an object of class \code{GSDesign}}
}
\description{
The class \code{GroupSequentialDesign} allows the implementation
and usage of group-sequential designs. These are two-stage designs
with a constant stage-two sample size function.
Therefore, the length of \code{n2_pivots} is required to be 1 in this
class.
\code{GroupSequentialDesign} is a subclass of \link{TwoStageDesign}.
}
\section{Slots}{

\describe{
\item{\code{n1}}{stage-one sample size}

\item{\code{c1f}}{early stopping for futility boundary}

\item{\code{c1e}}{early stopping for efficacy boundary}

\item{\code{n2_pivots}}{stage-two sample size upon continuation}

\item{\code{c2_pivots}}{vector of length order giving the values of c2 at the
pivot points of the numeric integration rule [TODO: these are not available during construction]}

\item{\code{x1_norm_pivots}}{normalized pivots for integration rule (in [-1, 1])}

\item{\code{weights}}{weights of conditional score values at x1_norm_pivots for
approximating the integral over x1.}

\item{\code{tunable}}{named logical vector indicating whether corresponding slot is considered a tunable parameter}
}}

