% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scores.R
\docType{class}
\name{IntegralScore-class}
\alias{IntegralScore-class}
\alias{evaluate,IntegralScore,Design-method}
\title{Score class obtained by integration of a \code{ConditionalScore}}
\usage{
\S4method{evaluate}{IntegralScore,Design}(s, design, specific = TRUE,
  ...)
}
\arguments{
\item{s}{an \code{IntegralScore}}

\item{design}{a \code{Design}}

\item{specific}{logical, flag for switching to design-specific implementation
(default).}
}
\description{
Score class obtained by integration of a \code{ConditionalScore}
}
\section{Methods (by generic)}{
\itemize{
\item \code{evaluate}: generic implementation of evaluating an integral
score. Uses adaptive Gaussian quadrature for integration and might be
more efficiently implemented by specific \code{Design}-classes
(cf. .evaluate).
}}

\section{Slots}{

\describe{
\item{\code{cs}}{the underlying \code{ConditionalScore}}
}}

