% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scores.R
\docType{class}
\name{IntegralScore-class}
\alias{IntegralScore-class}
\alias{evaluate,IntegralScore,TwoStageDesign-method}
\title{Unconditional score class obtained by integration of a \code{ConditionalScore}}
\usage{
\S4method{evaluate}{IntegralScore,TwoStageDesign}(s, design,
  optimization = FALSE, ...)
}
\arguments{
\item{s}{an \code{IntegralScore}}

\item{design}{a \code{TwoStageDesign}}

\item{...}{further optimal arguments}

\item{specific}{logical, flag for switching to design-specific implementation.}
}
\description{
Unconditional score class obtained by integration of a \code{ConditionalScore}
}
\section{Methods (by generic)}{
\itemize{
\item \code{evaluate}: generic implementation of evaluating an integral
score. Uses adaptive Gaussian quadrature for integration and might be
more efficiently implemented by specific \code{TwoStageDesign}-classes.
}}

\section{Slots}{

\describe{
\item{\code{cs}}{the underlying \code{ConditionalScore}}
}}

