% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scores.R
\docType{class}
\name{IntegralScore-class}
\alias{IntegralScore-class}
\alias{IntegralScore}
\alias{show,IntegralScore-method}
\title{Unconditional score class obtained by integration of a
\code{\link{ConditionalScore}}}
\usage{
\S4method{show}{IntegralScore}(object)
}
\arguments{
\item{object}{object of class \code{IntegralScore}}
}
\description{
Unconditional score class obtained by integration of a
\code{\link{ConditionalScore}}
}
\section{Slots}{

\describe{
\item{\code{cs}}{the underlying \code{\link{ConditionalScore}}}
}}

\examples{
expected(ConditionalPower(Normal(), PointMassPrior(.4, 1.0)))

}
\seealso{
The method \code{\link{expected}} creates a \code{IntegralScore}
   from a \code{\link{ConditionalScore}}.
}
