% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimize.R
\name{minimize}
\alias{minimize}
\title{Find optimal two-stage design by constraint minimization}
\usage{
minimize(objective, subject_to, initial_design, lower_boundary_design,
  upper_boundary_design, c2_monotone = F, opts = list(algorithm =
  "NLOPT_LN_COBYLA", xtol_rel = 1e-04, maxeval = 2500), ...)
}
\arguments{
\item{objective}{objective function}

\item{subject_to}{constraint collection}

\item{initial_design}{initial guess (x0 for nloptr)}

\item{lower_boundary_design}{design specifying the lower boundary}

\item{upper_boundary_design}{design specifying the upper boundary}

\item{c2_monotone}{should the c2-function be forced to be monotoneously decreasing?}

\item{opts}{options list passed to nloptr}

\item{...}{further optional arguments passed to \code{\link{nloptr}}}
}
\description{
\code{minimize} takes an unconditioonal score [TODO: need joint superclass for UnconditionalDesign and AffineScores...]
and a constraint set (or single constraint) and solves the corresponding
constraint minimization problem using \code{nloptr} (using COBYLA by default).
}
