% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDistribution.R
\docType{class}
\name{NestedModels-class}
\alias{NestedModels-class}
\alias{NestedModels}
\title{F-Distribution}
\usage{
NestedModels(p_inner, p_outer)
}
\arguments{
\item{p_inner}{number of independent variables in smaller model}

\item{p_outer}{number of independent variables in bigger model}

\item{x}{outcome}

\item{n}{sample size}

\item{theta}{distribution parameter}

\item{...}{further optional arguments}
}
\description{
Implements the F-distribution used for an ANOVA or for the comparison of the fit of two
nested regression models. In both cases, the test statistic follows a F-distribution.
\code{NestedModel} is used to compare the fit of two regression models, where one model contains
the independent variables of the smaller model as a subset. Then, one can use ANOVA to determine
whether more variance can be explained by adding more independent variables.
In the class \code{ANOVA}, the number of independent variables of the smaller model is set to \eqn{1}
in order to match the degrees of freedom and we obtain a one-way ANOVA.
}
\section{Slots}{

\describe{
\item{\code{p_inner}}{number of parameters in smaller model}

\item{\code{p_outer}}{number of parameters in bigger model}
}}

\examples{
model <- NestedModels(2, 4)

}
\seealso{
see \code{\link{probability_density_function}} and
   \code{\link{cumulative_distribution_function}} to evaluate the pdf
   and the cdf, respectively.
}
