% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDistribution.R
\docType{class}
\name{Normal-class}
\alias{Normal-class}
\alias{Normal}
\alias{probability_density_function,Normal,numeric,numeric,numeric-method}
\alias{cumulative_distribution_function,Normal,numeric,numeric,numeric-method}
\alias{quantile,Normal-method}
\alias{simulate,Normal,numeric-method}
\alias{show,Normal-method}
\title{Normal data distribution}
\usage{
Normal(two_armed = TRUE)


  \S4method{probability_density_function}{Normal,numeric,numeric,numeric}(dist,
  x, n, theta, ...)


  \S4method{cumulative_distribution_function}{Normal,numeric,numeric,numeric}(dist,
  x, n, theta, ...)

\S4method{quantile}{Normal}(x, probs, n, theta, ...)

\S4method{simulate}{Normal,numeric}(object, nsim, n, theta, seed = NULL,
  ...)

\S4method{show}{Normal}(object)
}
\arguments{
\item{two_armed}{s. slot}

\item{dist}{data distribution object}

\item{x}{outcome}

\item{n}{sample size}

\item{theta}{distribution parameter}

\item{...}{further optional arguments}

\item{probs}{vector of probabilities}

\item{object}{design to simulate from}

\item{nsim}{number of simulation runs}

\item{seed}{random seed}

\item{object}{object of class \code{Normal}}
}
\description{
Implements a normal data distribution for z-values given an observed z-value
and stage size.
Standard deviation is 1 and mean \eqn{\theta\sqrt n} where
\eqn{\theta} is the standardized effect size.
The option \code{two_armed} can be set to decide whether a one-arm or a
two-arm design should be computed.
See \code{\link{DataDistribution-class}} for more details.
}
