% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneStageDesign.R
\docType{class}
\name{OneStageDesign-class}
\alias{OneStageDesign-class}
\alias{OneStageDesign}
\alias{tunable_parameters,OneStageDesign-method}
\alias{update,OneStageDesign-method}
\alias{n2,OneStageDesign,numeric-method}
\alias{c2,OneStageDesign,numeric-method}
\alias{evaluate,SmoothnessN2,OneStageDesign-method}
\alias{TwoStageDesign,OneStageDesign-method}
\title{One-stage design}
\usage{
OneStageDesign(n1, c1f)

\S4method{tunable_parameters}{OneStageDesign}(x, ...)

\S4method{update}{OneStageDesign}(object, params, ...)

\S4method{n2}{OneStageDesign,numeric}(d, x1, ...)

\S4method{c2}{OneStageDesign,numeric}(d, x1, ...)

\S4method{evaluate}{SmoothnessN2,OneStageDesign}(s, design, ...)

\S4method{TwoStageDesign}{OneStageDesign}(d, ...)
}
\arguments{
\item{n1}{cf. slot}

\item{c1f}{cf. slot}

\item{x}{object to get parameters from}

\item{...}{further optional arguments}

\item{object}{object to update}

\item{params}{vector of design parameters (must be in same order as returned
by \code{as.numeric(OneStageDesign)})}

\item{d}{object of class \code{OneStageDesign}}

\item{x1}{stage-one outcome}

\item{s}{an object of class \code{SmoothnessN2}}

\item{design}{an object of class \code{OneStageDesign}}
}
\description{
[ToDo]
}
\section{Slots}{

\describe{
\item{\code{n1}}{sample size}

\item{\code{c1f}}{rejection boundary}

\item{\code{c1e}}{rejection boundary}

\item{\code{n2_pivots}}{constant 0}

\item{\code{c2_pivots}}{\code{+/- Inf}}

\item{\code{x1_norm_pivots}}{is ignored for OneStageDesign}

\item{\code{weights}}{is ignored for OneStageDesign}
}}

