% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneStageDesign.R
\name{OneStageDesign}
\alias{OneStageDesign}
\alias{TwoStageDesign,OneStageDesign-method}
\alias{plot,OneStageDesign-method}
\title{plot() is not defined for one stage designs}
\usage{
OneStageDesign(n, c)

\S4method{TwoStageDesign}{OneStageDesign}(n1, order = 5L, eps = 0.01, ...)

\S4method{plot}{OneStageDesign}(x, y, ...)
}
\arguments{
\item{n}{sample size (stage-one sample size)}

\item{c}{rejection boundary (\ifelse{html}{\out{c = c<sub>1</sub><sup>f</sup> = c<sub>1</sub><sup>e</sup>}}{\eqn{c = c_1^f = c_1^e}})}

\item{n1}{\code{OneStageDesign} object to convert, overloaded from
\code{\link{TwoStageDesign}}}

\item{order}{integer >= 2, default is 5; order of Gaussian quadrature
integration rule to use for new TwoStageDesign.}

\item{eps}{numeric > 0, default = .01; the single critical value c must be
split in a continuation interval [c1f, c1e]; this is given by c +/- eps.}

\item{...}{further optional arguments}

\item{x}{design to plot}

\item{y}{not used}
}
\description{
plot() is not defined for one stage designs
}
\examples{
design <- OneStageDesign(30, 1.96)
summary(design)
design <- TwoStageDesign(design)
summary(design)

}
