% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneStageDesign.R, R/regularization.R
\docType{class}
\name{OneStageDesign-class}
\alias{OneStageDesign-class}
\alias{OneStageDesign}
\alias{update,OneStageDesign-method}
\alias{n2,OneStageDesign,numeric-method}
\alias{c2,OneStageDesign,numeric-method}
\alias{TwoStageDesign,OneStageDesign-method}
\alias{plot,OneStageDesign-method}
\alias{evaluate,SmoothnessN2,OneStageDesign-method}
\title{One-stage design}
\usage{
OneStageDesign(n, c)

\S4method{update}{OneStageDesign}(object, params, ...)

\S4method{n2}{OneStageDesign,numeric}(d, x1, ...)

\S4method{c2}{OneStageDesign,numeric}(d, x1, ...)

\S4method{TwoStageDesign}{OneStageDesign}(d, ...)

\S4method{plot}{OneStageDesign}(x, y, ...)

\S4method{evaluate}{SmoothnessN2,OneStageDesign}(s, design, ...)
}
\arguments{
\item{n}{sample size (stage-one sample size)}

\item{c}{rejection boundary (c = c1f = c1e)}

\item{object}{object to update}

\item{params}{vector of design parameters (must be in same order as returned
by \code{as.numeric(OneStageDesign)})}

\item{...}{further optional arguments}

\item{d}{object of class \code{OneStageDesign}}

\item{x1}{stage-one outcome}

\item{x}{not used}

\item{y}{not used}

\item{s}{an object of class \code{SmoothnessN2}}

\item{design}{an object of class \code{OneStageDesign}}

\item{rounded}{logical that indicates whether rounded n1-value should be used}

\item{k}{number of points to use for plotting}
}
\description{
\code{OneStageDesign} allows the usage of a single-stage design.
In this case without a second stage, there only exist the first-stage sample
size \code{n1} and the first-stage rejection boundary \code{c1f}.
No other parameters have to be set.
\code{OneStageDesign} is a subclass of \link{TwoStageDesign}.
}
\section{Slots}{

\describe{
\item{\code{n1}}{sample size}

\item{\code{c1f}}{rejection boundary}

\item{\code{c1e}}{rejection boundary}

\item{\code{n2_pivots}}{constant 0}

\item{\code{c2_pivots}}{\code{+/- Inf}}

\item{\code{x1_norm_pivots}}{is ignored for OneStageDesign}

\item{\code{weights}}{is ignored for OneStageDesign}

\item{\code{tunable}}{is ignored for OneStageDesign}
}}

