% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointMassPrior.R
\docType{class}
\name{PointMassPrior-class}
\alias{PointMassPrior-class}
\alias{PointMassPrior}
\title{Point-mass prior}
\usage{
PointMassPrior(theta, mass)
}
\arguments{
\item{theta}{cf. slot 'theta'}

\item{mass}{cf. slot 'mass'}
}
\description{
\code{PointMassPrior} is a generic class for representing a univariate prior
over a discrete set of points with positive probability mass.
All methods for prior distributions [TODO link] are specialized to
the discrete case.
}
\section{Slots}{

\describe{
\item{\code{theta}}{numeric vector of pivot points (parameter values with positive
prior mass)}

\item{\code{mass}}{numeric vector of the same length as theta - corresponding
probability masses (must sum to one!)}
}}

