% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{postprocess}
\alias{postprocess}
\title{Post-process an optimal design}
\usage{
postprocess(results)
}
\arguments{
\item{results}{An object obtained by \code{\link{minimize}}.}
}
\value{
\item{design}{ The resulting optimal design}
        \item{nloptr_return}{ Output of the corresponding nloptr call}
}
\description{
\code{postprocess} takes an optimal design and rounds its sample sizes.
The corresponding decision boundaries are re-computed such that the
constraints which were specified in the underlying \code{minimize()} call
are fulfilled.
}
