% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoStageDesign.R
\name{print.TwoStageDesignSummary}
\alias{print.TwoStageDesignSummary}
\title{Print obejct of class TwoStageDesignSummary}
\usage{
\method{print}{TwoStageDesignSummary}(x, ..., rounded = T)
}
\arguments{
\item{x}{object to print}

\item{...}{unused}

\item{rounded}{should rounded n-values be used?}
}
\description{
Print obejct of class TwoStageDesignSummary
}
