% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularization.R
\docType{class}
\name{SmoothnessN2-class}
\alias{SmoothnessN2-class}
\alias{SmoothnessN2}
\alias{evaluate,SmoothnessN2,TwoStageDesign-method}
\title{Quadratic smoothness penalty term}
\usage{
SmoothnessN2(h = 0.1)

\S4method{evaluate}{SmoothnessN2,TwoStageDesign}(s, design,
  specific = TRUE, ...)
}
\arguments{
\item{h}{positive number, see slot \code{h}}

\item{s}{an object of class \code{SmoothnessN2}}

\item{design}{the design to compute the smoothness term for}

\item{specific}{should a specific implementation be used?}

\item{...}{further optional arguments}
}
\description{
\code{SmoothnessN2} is a generic class for implementing a smoothness penalty
via the average squared second derivative of the stage two sample size function
\code{n2} of a two-stage design.
The only parameter is the width used for the finite differences, \code{h}.
The generic implementation only evluates \code{n2} in the interior of the
continuation region of a design.
}
\section{Slots}{

\describe{
\item{\code{h}}{positive number giving the width of the central finite difference
interval for approximating the second derivative.}
}}

