% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoStageDesign.R
\docType{methods}
\name{summary,TwoStageDesign-method}
\alias{summary,TwoStageDesign-method}
\title{Summarize TwoStageDesign objects with optional set of scores}
\usage{
\S4method{summary}{TwoStageDesign}(object, ..., rounded = TRUE)
}
\arguments{
\item{object}{design object to summarize}

\item{...}{optinal additional named UnconditionalScores}

\item{rounded}{should rounded n-values be used?}
}
\description{
\code{summary} summarizes the first-stage of a \code{TwoStageDesign}
and objects of class \code{UnconditionalScore} that have to be
defined by the user.
}
\details{
The first stage sample size and the two continuation decsion boundaries
are printed.
Furthermore, the user can define unconditional scores and these will be
evaluated and listed.
}
\examples{
order  <- 5L
design <- TwoStageDesign(50, 0, 2, 50.0, 2.0, order)
pow    <- expected(ConditionalPower(dist = Normal(), prior = PointMassPrior(.4, 1)))
summary(design, "Power" = pow)

}
