% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoStageDesign.R
\docType{methods}
\name{summary,TwoStageDesign-method}
\alias{summary,TwoStageDesign-method}
\title{Summarize TwoStageDesign objects with optional set of scores}
\usage{
\S4method{summary}{TwoStageDesign}(object, ...)
}
\arguments{
\item{object}{design object to plot}

\item{...}{optinal additional named UnconditionalScores}
}
\description{
[TODO]
}
