% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoStageDesign.R
\docType{methods}
\name{plot,TwoStageDesign-method}
\alias{plot,TwoStageDesign-method}
\title{Plot TwoStageDesign with optional set of conditional scores}
\usage{
\S4method{plot}{TwoStageDesign}(x, ..., k = 100)
}
\arguments{
\item{x}{design object to plot}

\item{...}{optinal additional named ConditionalScores to plot}

\item{k}{number of points to use for plotting}
}
\description{
[TODO]
}
